native: fix build regressions blocking headless tests
Recent upstream-port commits left the native (Linux/Clang) build broken in
several places. This patch restores buildability and makes the engine run
through to gameplay in headless mode so the gameplay-telemetry test suite
can exercise the IK pipeline.

Compile fixes:
- src/system/obj/ObjPtr_p.h — remove duplicate definitions inside HX_NATIVE
  block (Node::RefOwner, operator<< for ObjPtrList/ObjPtrVec) that conflicted
  with the unconditional definitions and Object.h's inline copies.
- src/lazer/meta_ham/ChallengeSortNode.h — change Renumber(stlpmtx_std::...)
  to Renumber(std::vector<...>) to match base class and resolve undeclared
  namespace under native.
- src/lazer/meta_ham/OptionsPanel.cpp — cast switch expr to unsigned int and
  suffix case literals with U to clear narrowing warnings.
- src/system/char/CharEyes.cpp — drop local HX_NATIVE NormalizeScale def
  (already provided inline by Vec.h).
- src/lazer/meta_ham/Playlist.cpp — replace lvalue cast with const_cast in
  const member.
- src/system/obj/TypeProps.cpp — undeclared 'key' -> 'prop' (the actual
  parameter name).
- src/system/movie/Splash.cpp — restore DWORD return type to match header.
- src/system/utl/GlitchFinder.cpp — add explicit forward decl for
  GlitchFindScriptImpl so call sites before its definition compile.
- native/src/platform/NetworkSocket_Stub.cpp — InqBoundPort returns bool to
  match base virtual.

Link / stub fixes:
- src/system/char/CharClipSet.cpp — implement DrawShowing (declared in .h,
  never defined).
- src/system/os/HolmesKeyboard.cpp — implement HolmesInput::Handle stub.
- native/src/native_link_glue.cpp — add Ogg{Malloc,Calloc,Realloc,Free}
  stubs (VorbisMem.cpp is empty for native; Xbox link_glue.cpp is not
  compiled here).
- src/lazer/meta_ham/SongSortMgr.cpp — stub GetSetlistMode() returning
  false; was declared but not defined.

Runtime fixes (headless mode):
- src/lazer/meta_ham/ShellInput.cpp — recent upstream port assumed Kinect
  state is fully initialized. Renamed HandsUp()/RaisedMs() to the actual
  class methods (GetHandsUp/GetRaisedMs), declared sHasSkeleton in the
  HX_NATIVE block, added null guards on mHandsUpGestureFilter, and gated
  the Kinect/cursor/skeleton subsystem polls under #ifndef HX_NATIVE since
  their bodies dereference DataNode state that's not present in headless.
- src/lazer/meta_ham/VoiceInputPanel.cpp — null-guard TheSpeechMgr before
  ->Enabled() to avoid SIGSEGV on entering song_select_screen in headless
  (no SpeechMgr instance).
- native/src/audio/AudioDevice.cpp — add DC3_NO_AUDIO opt-out; MILO_HEADLESS
  implies it. Some hosts crash on the PipeWire/SPA audio thread shortly
  after init, which made headless test runs flaky.

With these in place, dc3-native boots through attract -> title -> main ->
choose_mode -> song_select -> loading -> game_screen and produces gameplay
telemetry samples on player 0 reliably.

+++ src/system/movie/Splash.cpp
@@ -515,7 +515,7 @@ void Splash::UpdateThread() {
     MILO_ASSERT(SetImmutableState(kTerminated), 0x257);
 }
 
-unsigned long Splash::ThreadStart(void *v) {
+DWORD Splash::ThreadStart(void *v) {
     static_cast<Splash *>(v)->UpdateThread();
     return 0;
 }
